A ggproto for drawing anchor points

This is a wrapper function used by explore_space_pca() and should be be called directly by the user

Usage

add_anchor(dt, anchor_size = 3, anchor_alpha = 0.5, anchor_color = NULL, ...)

Arguments

dt

A data object from the running the optimisation algorithm in guided tour

anchor_size

numeric; the size of the anchor points

anchor_alpha

numeric; the alpha of the anchor points

anchor_color

the variable to be coloured by

...

other aesthetics inherent from explore_space_pca()

Value

a wrapper for drawing anchor points in explore_space_pca()

See also