A ggproto for drawing interpolation path

This is a wrapper function used by explore_space_pca() and should be be called directly by the user

add_interp(
  dt,
  interp_size = 1.5,
  interp_alpha = NULL,
  interp_color = NULL,
  interp_group = NULL,
  ...
)

Arguments

dt
A data object from the running the optimisation algorithm in guided tour

interp_size
the size of the path

interp_alpha
an alpha value for the transparency of the path

interp_color
the colour of the path

interp_group
a group variable for path connection

...
other aesthetics inherent from explore_space_pca()

See also