Compute PCA for the projection bases

Compute PCA for the projection bases

compute_pca(dt, group = NULL, random = TRUE)

Arguments

dt
A data object from the running the optimisation algorithm in guided tour

group
The grouping variable, useful when there are multiple algorithms in the data object

random
Boolean, if the random data from the high dimensional sphere should be bounded

Examples

dplyr::bind_rows(holes_1d_geo, holes_1d_better) %>% compute_pca(group = method)

#> signs in all the bases will be flipped in group search_geodesic

#> $pca_summary
#> Standard deviations (1, .., p=5):
#> [1] 1.0422652 1.0156862 0.9891546 0.9851991 0.9659300
#> 
#> Rotation (n x k) = (5 x 5):
#>           PC1        PC2        PC3        PC4         PC5
#> V1  0.5519691 -0.1961065  0.5944625 0.10779747 -0.54024668
#> V2 -0.3658798  0.5052715  0.1696970 0.72896372 -0.22504985
#> V3 -0.4422308 -0.4151056  0.6624928 0.04648432  0.43710616
#> V4  0.1147173 -0.6607185 -0.4045595 0.62076143  0.03574785
#> V5  0.5939141  0.3120598  0.1235001 0.26360206  0.68201641
#> 
#> $aug
#> # A tibble: 2,987 x 14
#>    basis index_val info  method alpha tries  loop    id row_num   PC1    PC2
#>    <lis>     <dbl> <chr> <chr>  <dbl> <dbl> <dbl> <dbl>   <int> <dbl>  <dbl>
#>  1 <dbl…     0.749 new_… searc…   0.5     1     1     1       1 -1.84 -0.635
#>  2 <dbl…     0.749 dire… searc…  NA       2     1     2       2 -1.84 -0.646
#>  3 <dbl…     0.749 dire… searc…  NA       2     1     3       3 -1.83 -0.624
#>  4 <dbl…     0.749 dire… searc…  NA       2     1     4       4 -1.83 -0.625
#>  5 <dbl…     0.749 dire… searc…  NA       2     1     5       5 -1.84 -0.645
#>  6 <dbl…     0.749 dire… searc…  NA       2     1     6       6 -1.83 -0.614
#>  7 <dbl…     0.749 dire… searc…  NA       2     1     7       7 -1.84 -0.655
#>  8 <dbl…     0.749 dire… searc…  NA       2     1     8       8 -1.83 -0.645
#>  9 <dbl…     0.749 dire… searc…  NA       2     1     9       9 -1.84 -0.625
#> 10 <dbl…     0.749 best… searc…  NA       2     1    10      10 -1.84 -0.653
#> # … with 2,977 more rows, and 3 more variables: PC3 <dbl>, PC4 <dbl>, PC5 <dbl>
#>