ferrn: Facilitate Exploration of touRR optimisatioN

Diagnostic plots for optimisation, with a focus on projection pursuit. These show paths the optimiser takes in the high-dimensional space in multiple ways: by reducing the dimension using principal component analysis, and also using the tour to show the path on the high-dimensional space. Several botanical colour palettes are included, reflecting the name of the package. A paper describing the methodology can be found at https://journal.r-project.org/archive/2021/RJ-2021-105/index.html.

See also

Author

Maintainer: H. Sherry Zhang huizezhangsh@gmail.com (ORCID)

Authors:

Dianne Cook dicook@monash.edu (ORCID)

Ursula Laa ursula.laa@boku.ac.at (ORCID)

Nicolas Langrené nicolaslangrene@uic.edu.cn (ORCID)

Patricia Menéndez patricia.menendez@unimelb.edu.au (ORCID)