Flip the sign of a group of bases

Flip the sign of a group of bases

flip_sign(dt, group = NULL)

Arguments

dt
A data object from the running the optimisation algorithm in guided tour

group
The grouping variable, useful when there are multiple algorithms in the data object

Examples

dplyr::bind_rows(holes_1d_geo, holes_1d_better) %>% flip_sign(group = method)

#> signs in all the bases will be fliped in group search_geodesic

#> $basis
#>                   V1          V2            V3            V4           V5
#>   [1,] -0.3410058690  0.11290652  1.451996e-01 -0.0357833010 -0.921196912
#>   [2,] -0.3377096820  0.11686570  1.528816e-01 -0.0320019378 -0.920813590
#>   [3,] -0.3442679556  0.10893605  1.375030e-01 -0.0395610858 -0.921488116
#>   [4,] -0.3346916725  0.11796831  1.448883e-01 -0.0414566863 -0.922688286
#>   [5,] -0.3472859651  0.10783344  1.454963e-01 -0.0301063373 -0.919613419
#>   [6,] -0.3403675482  0.11618411  1.402738e-01 -0.0438160889 -0.921441590
#>   [7,] -0.3416100894  0.10961764  1.501109e-01 -0.0277469347 -0.920860116
#>   [8,] -0.3360575570  0.10476594  1.449172e-01 -0.0369989237 -0.923969430
#>   [9,] -0.3459200806  0.12103581  1.454675e-01 -0.0345641000 -0.918332275
#>  [10,] -0.3423378752  0.11228816  1.502015e-01 -0.0273005279 -0.920266447
#>  [11,] -0.3396397624  0.11351359  1.401831e-01 -0.0442624957 -0.922035259
#>  [12,] -0.3661418429  0.07677815  3.419815e-01  0.3233736296 -0.799076601
#>  [13,] -0.3410058690  0.11290652  1.451996e-01 -0.0357833010 -0.921196912
#>  [14,] -0.3471626395  0.10979345  1.693865e-01  0.0055002841 -0.915806457
#>  [15,] -0.3524980253  0.10642061  1.931726e-01  0.0467708555 -0.908249208
#>  [16,] -0.3569994026  0.10279598  2.165017e-01  0.0879307674 -0.898543047
#>  [17,] -0.3606561214  0.09892814  2.393185e-01  0.1288826355 -0.886710938
#>  [18,] -0.3634595298  0.09482623  2.615692e-01  0.1695295681 -0.872780876
#>  [19,] -0.3654029949  0.09049997  2.832009e-01  0.2097753946 -0.856785820
#>  [20,] -0.3664819186  0.08595958  3.041626e-01  0.2495248937 -0.838763612
#>  [21,] -0.3666937481  0.08121581  3.244046e-01  0.2886840183 -0.818756895
#>  [22,] -0.3691722751  0.08139709  3.210720e-01  0.2801486545 -0.821897691
#>  [23,] -0.3641785521  0.08102641  3.277048e-01  0.2971905140 -0.815534224
#>  [24,] -0.3675015992  0.08989653  3.266889e-01  0.2915902320 -0.815543196
#>  [25,] -0.3658492280  0.07252697  3.220880e-01  0.2857489365 -0.821888719
#>  [26,] -0.3726444459  0.08680691  3.251258e-01  0.2831931927 -0.817126353
#>  [27,] -0.3607063813  0.07561659  3.236510e-01  0.2941459758 -0.820305562
#>  [28,] -0.3666612324  0.07350869  3.203883e-01  0.2935966936 -0.819334072
#>  [29,] -0.3666895948  0.08891481  3.283885e-01  0.2837424749 -0.818097843
#>  [30,] -0.3602447291  0.08291962  3.231053e-01  0.2833622527 -0.823806334
#>  [31,] -0.3731060982  0.07950388  3.256716e-01  0.2939769158 -0.813625581
#>  [32,] -0.2614694750  0.59989789  5.119944e-01 -0.1417536234 -0.538074231
#>  [33,] -0.3666937481  0.08121581  3.244046e-01  0.2886840183 -0.818756895
#>  [34,] -0.3662665813  0.12137017  3.448705e-01  0.2625344190 -0.814406598
#>  [35,] -0.3648379933  0.16119269  3.643934e-01  0.2356670159 -0.807829607
#>  [36,] -0.3624118900  0.20057448  3.829200e-01  0.2081552681 -0.799043902
#>  [37,] -0.3589949049  0.23940788  4.003997e-01  0.1800743963 -0.788073506
#>  [38,] -0.3545963802  0.27758670  4.167846e-01  0.1515011774 -0.774948414
#>  [39,] -0.3492283423  0.31500657  4.320299e-01  0.1225137343 -0.759704509
#>  [40,] -0.3429054680  0.35156517  4.460941e-01  0.0931913224 -0.742383473
#>  [41,] -0.3356450449  0.38716254  4.589385e-01  0.0636141131 -0.723032662
#>  [42,] -0.3274669240  0.42170135  4.705282e-01  0.0338629745 -0.701704985
#>  [43,] -0.3183934653  0.45508719  4.808314e-01  0.0040192501 -0.678458753
#>  [44,] -0.3084494768  0.48722875  4.898199e-01 -0.0258354636 -0.653357526
#>  [45,] -0.2976621467  0.51803816  4.974692e-01 -0.0556195396 -0.626469932
#>  [46,] -0.2860609691  0.54743119  5.037583e-01 -0.0852515445 -0.597869488
#>  [47,] -0.2736776630  0.57532747  5.086701e-01 -0.1146504604 -0.567634389
#>  [48,] -0.2691516156  0.57797777  5.146680e-01 -0.1120040083 -0.562201942
#>  [49,] -0.2781763429  0.57261965  5.026214e-01 -0.1172854476 -0.573010074
#>  [50,] -0.2784690616  0.57699532  5.089136e-01 -0.1211746737 -0.562009858
#>  [51,] -0.2688588969  0.57360210  5.083759e-01 -0.1081147822 -0.573202158
#>  [52,] -0.2751680732  0.56991691  5.061564e-01 -0.1199215713 -0.573499562
#>  [53,] -0.2721598853  0.58068050  5.111330e-01 -0.1093678846 -0.561712453
#>  [54,] -0.2782352310  0.57285639  5.148937e-01 -0.1161681277 -0.561969937
#>  [55,] -0.2690927274  0.57774103  5.023958e-01 -0.1131213282 -0.573242078
#>  [56,] -0.2696085869  0.57116845  5.073551e-01 -0.1196846340 -0.573885163
#>  [57,] -0.2777193716  0.57942896  5.099343e-01 -0.1096048219 -0.561326853
#>  [58,] -0.1725095879  0.75947158  5.660485e-01  0.1435534263 -0.228964854
#>  [59,] -0.2736776630  0.57532747  5.086701e-01 -0.1146504604 -0.567634389
#>  [60,] -0.2659779179  0.60101015  5.202530e-01 -0.0886531257 -0.538070569
#>  [61,] -0.2576390464  0.62524864  5.305858e-01 -0.0624427637 -0.507213803
#>  [62,] -0.2486810860  0.64798470  5.396436e-01 -0.0360823560 -0.475138237
#>  [63,] -0.2391255623  0.66916370  5.474047e-01 -0.0096352449 -0.441920947
#>  [64,] -0.2289954365  0.68873474  5.538505e-01  0.0168350190 -0.407641751
#>  [65,] -0.2183150504  0.70665081  5.589653e-01  0.0432648295 -0.372383021
#>  [66,] -0.2071100685  0.72286884  5.627370e-01  0.0695906776 -0.336229480
#>  [67,] -0.1954074154  0.73734986  5.651565e-01  0.0957493042 -0.299268003
#>  [68,] -0.1832352119  0.75005909  5.662179e-01  0.1216778518 -0.261587405
#>  [69,] -0.1878037725  0.74871499  5.703785e-01  0.1183170146 -0.254607748
#>  [70,] -0.1786483279  0.75132819  5.620008e-01  0.1250265214 -0.268540904
#>  [71,] -0.1785299982  0.75197795  5.634433e-01  0.1298292788 -0.261404383
#>  [72,] -0.1879221023  0.74806522  5.689360e-01  0.1135142572 -0.261744269
#>  [73,] -0.1855743176  0.75338593  5.656101e-01  0.1177837041 -0.253345678
#>  [74,] -0.1808777829  0.74665725  5.667691e-01  0.1255598319 -0.269802974
#>  [75,] -0.1765245827  0.74763327  5.727521e-01  0.1212873355 -0.259090628
#>  [76,] -0.1899275178  0.75240991  5.596271e-01  0.1220562004 -0.264058024
#>  [77,] -0.1871358696  0.75472516  5.600910e-01  0.1243481714 -0.257304614
#>  [78,] -0.1793162308  0.74531801  5.722882e-01  0.1189953646 -0.265844038
#>  [79,] -0.3361364768  0.88035970  2.414241e-01  0.2267240298 -0.047850507
#>  [80,] -0.1832352119  0.75005909  5.662179e-01  0.1216778518 -0.261587405
#>  [81,] -0.2020894978  0.77211060  5.357945e-01  0.1345887147 -0.240447993
#>  [82,] -0.2204626275  0.79232378  5.040953e-01  0.1471791346 -0.218736096
#>  [83,] -0.2383108565  0.81065052  4.711960e-01  0.1594191347 -0.196503409
#>  [84,] -0.2555916897  0.82704717  4.371747e-01  0.1712795728 -0.173802867
#>  [85,] -0.2722639831  0.84147470  4.021126e-01  0.1827322104 -0.150688515
#>  [86,] -0.2882880416  0.85389876  3.660931e-01  0.1937497798 -0.127215389
#>  [87,] -0.3036257133  0.86428977  3.292020e-01  0.2043060491 -0.103439375
#>  [88,] -0.3182404807  0.87262298  2.915271e-01  0.2143758850 -0.079417081
#>  [89,] -0.3320975474  0.87887856  2.531581e-01  0.2239353120 -0.055205703
#>  [90,] -0.3349009081  0.87712399  2.501349e-01  0.2278807606 -0.063227917
#>  [91,] -0.3292609773  0.88054525  2.561559e-01  0.2199674700 -0.047177968
#>  [92,] -0.3395129774  0.87789579  2.512833e-01  0.2180770226 -0.057697784
#>  [93,] -0.3246489080  0.87977344  2.550075e-01  0.2297712080 -0.052708101
#>  [94,] -0.3280359484  0.88226151  2.450554e-01  0.2259845375 -0.053720285
#>  [95,] -0.3361259370  0.87540773  2.612354e-01  0.2218636931 -0.056685600
#>  [96,] -0.3331236019  0.87530718  2.595910e-01  0.2297514418 -0.051891947
#>  [97,] -0.3310382835  0.88236205  2.466998e-01  0.2180967888 -0.058513938
#>  [98,] -0.3366733391  0.87617803  2.494380e-01  0.2310913454 -0.057798089
#>  [99,] -0.3274885463  0.88149120  2.568528e-01  0.2167568852 -0.052607796
#> [100,] -0.2684603424  0.94701894 -7.976283e-03 -0.0117861650 -0.175731705
#> [101,] -0.3320975474  0.87887856  2.531581e-01  0.2239353120 -0.055205703
#> [102,] -0.3266724443  0.89489415  2.217221e-01  0.1955207085 -0.071137399
#> [103,] -0.3204844440  0.90881984  1.897684e-01  0.1666494942 -0.086902964
#> [104,] -0.3135479978  0.92062312  1.573714e-01  0.1373890937 -0.102465580
#> [105,] -0.3058793047  0.93027640  1.246070e-01  0.1078078404 -0.117788901
#> [106,] -0.2974962739  0.93775716  9.155156e-02  0.0779748172 -0.132837144
#> [107,] -0.2884184828  0.94304793  5.828232e-02  0.0479596949 -0.147575164
#> [108,] -0.2786671313  0.94613633  2.487696e-02  0.0178325695 -0.161968543
#> [109,] -0.2778853461  0.94792694  2.074649e-02  0.0152135930 -0.153467846
#> [110,] -0.2794210500  0.94425111  2.900494e-02  0.0204497628 -0.170453044
#> [111,] -0.2823547686  0.94589905  1.647728e-02  0.0170001919 -0.158083100
#> [112,] -0.2749516275  0.94627900  3.327415e-02  0.0186631638 -0.165837790
#> [113,] -0.2858981815  0.94325797  2.625987e-02  0.0134995639 -0.166297367
#> [114,] -0.2714082146  0.94892008  2.349156e-02  0.0221637919 -0.157623523
#> [115,] -0.2812368132  0.94600630  1.565935e-02  0.0161605624 -0.159598090
#> [116,] -0.2760695829  0.94617175  3.409208e-02  0.0195027934 -0.164322800
#> [117,] -0.2718828075  0.94819581  1.788985e-02  0.0187362807 -0.162275539
#> [118,] -0.2854235886  0.94398224  3.186158e-02  0.0169270751 -0.161645351
#> [119,] -0.0885995836  0.98687457 -1.053411e-02  0.1256199903 -0.048346109
#> [120,] -0.2786671313  0.94613633  2.487696e-02  0.0178325695 -0.161968543
#> [121,] -0.2438436827  0.95854665  1.825610e-02  0.0384457967 -0.141128363
#> [122,] -0.2084667716  0.96878131  1.159381e-02  0.0589717619 -0.119967857
#> [123,] -0.1726166947  0.97681709  4.905204e-03  0.0793638764 -0.098535056
#> [124,] -0.1363748223  0.98263575 -1.794537e-03  0.0995758552 -0.076878605
#> [125,] -0.0998234141  0.98622408 -8.490205e-03  0.1195618226 -0.055047659
#> [126,] -0.1050240605  0.98590698 -1.518086e-02  0.1154198239 -0.058353849
#> [127,] -0.0946127855  0.98644255 -1.798702e-03  0.1236918651 -0.051735964
#> [128,] -0.0983941180  0.98521176 -9.091517e-03  0.1292887063 -0.053648499
#> [129,] -0.1012427281  0.98713777 -7.888043e-03  0.1098229827 -0.056441315
#> [130,] -0.0931037397  0.98615808 -1.099571e-02  0.1261569100 -0.052796426
#> [131,] -0.1065331064  0.98619145 -5.983850e-03  0.1129547790 -0.057293387
#> [132,] -0.1061046497  0.98620612 -4.014161e-03  0.1170597803 -0.049195331
#> [133,] -0.0935321963  0.98614341 -1.296540e-02  0.1220519087 -0.060894482
#> [134,] -0.1012255241  0.98505445 -1.053902e-02  0.1287624840 -0.052252052
#> [135,] -0.0984113219  0.98729508 -6.440542e-03  0.1103492050 -0.057837761
#> [136,] -0.0747021092  0.99100609  2.614044e-02 -0.0371864182 -0.101293491
#> [137,] -0.0998234141  0.98622408 -8.490205e-03  0.1195618226 -0.055047659
#> [138,] -0.0930489410  0.99042257  1.222303e-03  0.0758350014 -0.068209869
#> [139,] -0.0860657469  0.99239941  1.093207e-02  0.0319380722 -0.081219075
#> [140,] -0.0788894958  0.99215018  2.061731e-02 -0.0120304983 -0.094046097
#> [141,] -0.0743352760  0.99229550  1.258813e-02 -0.0140748543 -0.097300317
#> [142,] -0.0834358268  0.99190564  2.864443e-02 -0.0099849391 -0.090782472
#> [143,] -0.0859242903  0.99116020  2.454864e-02 -0.0075354858 -0.097770389
#> [144,] -0.0718468124  0.99304094  1.668393e-02 -0.0165243077 -0.090312400
#> [145,] -0.0708421693  0.99302213  1.513022e-02 -0.0129220551 -0.092154975
#> [146,] -0.0869289335  0.99117901  2.610234e-02 -0.0111377383 -0.095927813
#> [147,] -0.0734209180  0.99181085  2.079753e-02 -0.0165874161 -0.101058992
#> [148,] -0.0843501847  0.99239029  2.043503e-02 -0.0074723774 -0.087023797
#> [149,] -0.0743648676  0.99173726  1.775349e-02 -0.0151076227 -0.101900156
#> [150,] -0.0834062351  0.99246388  2.347907e-02 -0.0089521708 -0.086182633
#> [151,] -0.0053565069  0.99655135 -2.033673e-02 -0.0586456621 -0.054807086
#> [152,] -0.0788894958  0.99215018  2.061731e-02 -0.0120304983 -0.094046097
#> [153,] -0.0455724613  0.99548337  2.030136e-03 -0.0332346835 -0.076337089
#> [154,] -0.0121538073  0.99659678 -1.656157e-02 -0.0543647605 -0.058457861
#> [155,] -0.0176166584  0.99591344 -1.993250e-02 -0.0556238706 -0.065990517
#> [156,] -0.0066897409  0.99718047 -1.318898e-02 -0.0531002140 -0.050919360
#> [157,] -0.0212052376  0.99671386 -1.666111e-02 -0.0537676695 -0.054251837
#> [158,] -0.0031011617  0.99638005 -1.646036e-02 -0.0549564151 -0.062658040
#> [159,] -0.0134138528  0.99720484 -1.924263e-02 -0.0506294814 -0.049688520
#> [160,] -0.0108925465  0.99588907 -1.387885e-02 -0.0580946032 -0.067221356
#> [161,] -0.0185634476  0.99657747 -8.893204e-03 -0.0543439801 -0.058791130
#> [162,] -0.0057429517  0.99651644 -2.422828e-02 -0.0543801044 -0.058118746
#> [163,] -0.0122571531  0.99649855 -9.922287e-03 -0.0504175567 -0.064807537
#> [164,] -0.0120492462  0.99659536 -2.319919e-02 -0.0583065278 -0.052102340
#> [165,]  0.0137742881  0.99848000 -5.498903e-04 -0.0483184367 -0.022649119
#> [166,] -0.0121538073  0.99659678 -1.656157e-02 -0.0543647605 -0.058457861
#> [167,]  0.0136152436  0.99847530 -6.481690e-04 -0.0483558790 -0.022869058
#> [168,]  0.0057550275  0.99872094 -5.626199e-03 -0.0447994851 -0.022016353
#> [169,]  0.0214740982  0.99812981  4.329926e-03 -0.0519074373 -0.023719477
#> [170,]  0.0163274971  0.99875440 -7.222456e-03 -0.0426342541 -0.018793926
#> [171,]  0.0109016285  0.99809635  5.926183e-03 -0.0540726683 -0.026941904
#> [172,]  0.0100148200  0.99888281 -4.888696e-03 -0.0400670775 -0.022440395
#> [173,]  0.0172143056  0.99796794  3.592423e-03 -0.0566398450 -0.023295435
#> [174,]  0.0051749937  0.99862006  4.291127e-03 -0.0464155940 -0.023630008
#> [175,]  0.0220541320  0.99823069 -5.587400e-03 -0.0502913284 -0.022105822
#> [176,]  0.0124228131  0.99857032 -1.174348e-03 -0.0502874647 -0.013144843
#> [177,]  0.0148063126  0.99828043 -1.219250e-04 -0.0464194578 -0.032590987
#> [178,] -0.0032809561  0.99866495  9.238876e-03 -0.0444671651 -0.024389957
#> [179,]  0.0166730418  0.99807765 -4.594764e-03 -0.0562730201 -0.019371319
#> [180,]  0.0105560839  0.99877310  3.298490e-03 -0.0404339024 -0.026364511
#> [181,]  0.0120979747  0.99851305  4.061499e-03 -0.0435946815 -0.030138560
#> [182,]  0.0151311510  0.99833770 -5.357773e-03 -0.0531122409 -0.015597270
#> [183,]  0.0041146448  0.99865970 -2.818923e-04 -0.0455640101 -0.024201384
#> [184,]  0.0231144809  0.99819104 -1.014381e-03 -0.0511429124 -0.021534445
#> [185,]  0.0137757885  0.99861652 -9.384255e-03 -0.0463345127 -0.018447710
#> [186,]  0.0134533371  0.99823423  8.087982e-03 -0.0503724097 -0.027288120
#> [187,]  0.0194470815  0.99802033 -8.719539e-04 -0.0561412623 -0.020606412
#> [188,]  0.0077820442  0.99883042 -4.243193e-04 -0.0405656602 -0.025129418
#> [189,]  0.0133652228  0.99806228  1.281188e-02 -0.0514607875 -0.029676672
#> [190,]  0.0108961779  0.99896845  3.836849e-04 -0.0404706218 -0.017472703
#> [191,]  0.0163329478  0.99788230 -1.679958e-03 -0.0562363007 -0.028263126
#> [192,]  0.0115281522  0.99880776  4.427841e-03 -0.0402195428 -0.024757667
#> [193,]  0.0157009735  0.99804299 -5.724114e-03 -0.0564873796 -0.020978163
#> [194,]  0.0172679477  0.99834208 -7.418757e-03 -0.0515652361 -0.017346528
#> [195,]  0.0099611780  0.99850867  6.122484e-03 -0.0451416864 -0.028389301
#> [196,]  0.0084039050  0.99891205 -3.064475e-03 -0.0423719453 -0.017300902
#> [197,]  0.0188252207  0.99793870  1.768202e-03 -0.0543349771 -0.028434928
#> [198,]  0.0191679594  0.99808189  5.263432e-03 -0.0516979073 -0.027653990
#> [199,]  0.0080611662  0.99876885 -6.559705e-03 -0.0450090151 -0.018081840
#> [200,]  0.0084984453  0.99849405  8.832106e-03 -0.0438540168 -0.030597853
#> [201,]  0.0145424027  0.99810465  3.753826e-03 -0.0564989307 -0.019220043
#> [202,]  0.0126867230  0.99874610 -5.050099e-03 -0.0402079918 -0.026515787
#> [203,]  0.0224144115  0.99829486 -3.958381e-03 -0.0499483526 -0.019861107
#> [204,]  0.0048147142  0.99855589  2.662108e-03 -0.0467585698 -0.025874723
#> [205,]  0.0148038652  0.99817372  5.033311e-03 -0.0552981556 -0.018623946
#> [206,]  0.0124252605  0.99867703 -6.329584e-03 -0.0414087669 -0.027111884
#> [207,]  0.0112674512  0.99872246 -1.187965e-03 -0.0403088279 -0.028289348
#> [208,]  0.0159616744  0.99812829 -1.083083e-04 -0.0563980946 -0.017446482
#> [209,]  0.0080088621  0.99891994 -3.119417e-03 -0.0416619863 -0.018691178
#> [210,]  0.0192202636  0.99793081  1.823144e-03 -0.0550449362 -0.027044651
#> [211,] -0.0010678429  0.99855408  4.874554e-03 -0.0456883185 -0.027882272
#> [212,]  0.0085255876  0.99866880 -8.340551e-03 -0.0445102450 -0.023178054
#> [213,]  0.0187035380  0.99818195  7.044278e-03 -0.0521966775 -0.022557775
#> [214,]  0.0066398121  0.99887965  4.666704e-04 -0.0413363121 -0.022055527
#> [215,]  0.0205893136  0.99797110 -1.762944e-03 -0.0553706103 -0.023680303
#> [216,]  0.0181960784  0.99860849  4.768277e-03 -0.0465796727 -0.016049631
#> [217,]  0.0090330473  0.99824226 -6.064550e-03 -0.0501272498 -0.029686199
#> [218,]  0.0106652477  0.99859857  4.745317e-03 -0.0493775285 -0.015049337
#> [219,]  0.0165638779  0.99825217 -6.041590e-03 -0.0473293940 -0.030686493
#> [220,]  0.0124479639  0.99877343  6.229269e-04 -0.0460368201 -0.013300672
#> [221,]  0.0147811618  0.99807732 -1.919200e-03 -0.0506701024 -0.032435158
#> [222,]  0.0085066122  0.99862551  8.690420e-03 -0.0501217511 -0.009328157
#> [223,]  0.0165336760  0.99806414  5.698790e-03 -0.0554542180 -0.022067458
#> [224,]  0.0106954496  0.99878661 -6.995063e-03 -0.0412527045 -0.023668372
#> [225,]  0.0087851240  0.99866598 -4.452626e-03 -0.0423038847 -0.027922000
#> [226,]  0.0184440016  0.99818477  3.156353e-03 -0.0544030378 -0.017813830
#> [227,]  0.0173122737  0.99873165 -4.962067e-03 -0.0443190898 -0.015702467
#> [228,]  0.0099168520  0.99811910  3.665794e-03 -0.0523878327 -0.030033363
#> [229,]  0.0134368127  0.99804037 -5.202466e-03 -0.0525961564 -0.030682930
#> [230,]  0.0137923129  0.99881037  3.906193e-03 -0.0441107660 -0.015052900
#> [231,]  0.0154801951  0.99806110  3.444475e-03 -0.0514459821 -0.031238711
#> [232,]  0.0117489306  0.99878965 -4.740748e-03 -0.0452609404 -0.014497119
#> [233,]  0.0169893038  0.99799091  6.689955e-03 -0.0565622781 -0.021942071
#> [234,]  0.0167013945  0.99883850 -4.668999e-03 -0.0401802988 -0.020160602
#> [235,]  0.0105277312  0.99801224  3.372726e-03 -0.0565266236 -0.025575228
#> [236,]  0.0067909041  0.99887266 -3.283927e-03 -0.0417214049 -0.021350441
#> [237,]  0.0204382216  0.99797809  1.987654e-03 -0.0549855176 -0.024385388
#> [238,]  0.0154363736  0.99836993  8.803554e-03 -0.0478528952 -0.025530542
#> [239,]  0.0117927521  0.99848082 -1.009983e-02 -0.0488540272 -0.020205288
#> [240,]  0.0202817966  0.99831249 -7.728626e-03 -0.0476630096 -0.025086442
#> [241,]  0.0069473291  0.99853826  6.432353e-03 -0.0490439129 -0.020649388
#> [242,]  0.0162897600  0.99831973  8.128920e-03 -0.0480263218 -0.026828355
#> [243,]  0.0109393657  0.99853102 -9.425193e-03 -0.0486806007 -0.018907475
#> [244,] -0.0025796752  0.99845499  1.654649e-02 -0.0500183318 -0.017474778
#> [245,]  0.0066246158  0.99850977  2.197153e-03 -0.0457661805 -0.028896222
#> [246,]  0.0206045098  0.99834098 -3.493426e-03 -0.0509407420 -0.016839607
#> [247,]  0.0082127291  0.99831745  5.151098e-03 -0.0536274727 -0.019808684
#> [248,]  0.0190163966  0.99853330 -6.447371e-03 -0.0430794497 -0.025927146
#> [249,]  0.0227132174  0.99813176 -3.091077e-03 -0.0516803043 -0.023166415
#> [250,]  0.0045159082  0.99871898  1.794804e-03 -0.0450266182 -0.022569415
#> [251,]  0.0176282886  0.99802693 -1.397749e-03 -0.0513706091 -0.031473890
#> [252,]  0.0096008371  0.99882382  1.014757e-04 -0.0453363133 -0.014261940
#> [253,]  0.0175702612  0.99838114  3.863133e-03 -0.0448193631 -0.030044666
#> [254,]  0.0096588645  0.99846961 -5.159406e-03 -0.0518875594 -0.015691164
#> [255,]  0.0012983116  0.99796985  1.257126e-02 -0.0603653403 -0.015889647
#> [256,]  0.0232821546  0.99834272  3.026397e-04 -0.0465972223 -0.024461387
#> [257,]  0.0039469711  0.99850803 -1.598913e-03 -0.0501097001 -0.021274443
#> [258,]  0.0156498644  0.99801406  2.240398e-03 -0.0574075208 -0.020551579
#> [259,]  0.0115792613  0.99883668 -3.536671e-03 -0.0392994017 -0.025184251
#> [260,]  0.0090189801  0.99823909 -7.007850e-03 -0.0506956141 -0.028604721
#> [261,]  0.0182101456  0.99861166  5.711577e-03 -0.0460113084 -0.017131109
#> [262,]  0.0168543264  0.99799624 -1.286563e-03 -0.0522283094 -0.031464100
#> [263,]  0.0103747992  0.99885451 -9.710246e-06 -0.0444786131 -0.014271729
#> [264,]  0.0208727830  0.99800620  4.275179e-04 -0.0524096667 -0.028302007
#> [265,]  0.0063563427  0.99884455 -1.723791e-03 -0.0442972558 -0.017433822
#> [266,]  0.0011069534  0.99849867 -1.878152e-03 -0.0506239126 -0.020805639
#> [267,]  0.0164802350  0.99861871 -8.129696e-03 -0.0426311545 -0.024608229
#> [268,]  0.0107488907  0.99823204  6.833422e-03 -0.0540757679 -0.021127601
#> [269,]  0.0103818658  0.99841700 -5.579032e-03 -0.0526151229 -0.016007100
#> [270,]  0.0168472599  0.99843375  4.282759e-03 -0.0440917996 -0.029728729
#> [271,]  0.0097606619  0.99871812  5.237224e-03 -0.0419862203 -0.026011032
#> [272,]  0.0174684638  0.99813263 -6.533497e-03 -0.0547207022 -0.019724798
#> [273,]  0.0112425332  0.99856781 -4.576938e-03 -0.0501428054 -0.014166615
#> [274,]  0.0159865924  0.99828294  3.280665e-03 -0.0465641171 -0.031569215
#> [275,]  0.0061833701  0.99850076  4.968978e-03 -0.0503889553 -0.019855749
#> [276,]  0.0210457555  0.99834998 -6.265252e-03 -0.0463179671 -0.025880080
#> [277,] -0.0036697073  0.99838061  1.241463e-02 -0.0530763521 -0.015858071
#> [278,]  0.0177792493  0.99878794  2.358342e-03 -0.0428380973 -0.016305717
#> [279,]  0.0094498764  0.99806280 -3.654615e-03 -0.0538688252 -0.029430113
#> [280,]  0.0093614337  0.99864057 -1.593079e-04 -0.0493602002 -0.013889676
#> [281,]  0.0178676920  0.99821018 -1.136965e-03 -0.0473467223 -0.031846154
#> [282,]  0.0122735509  0.99815891 -8.435916e-03 -0.0517062198 -0.027990108
#> [283,]  0.0149555748  0.99869184  7.139643e-03 -0.0450007027 -0.017745722
#> [284,]  0.0074464670  0.99873844 -5.652662e-03 -0.0462501545 -0.017177214
#> [285,]  0.0197826587  0.99811231  4.356389e-03 -0.0504567680 -0.028558616
#> [286,]  0.0104806955  0.99889085 -8.808696e-04 -0.0389716834 -0.024241579
#> [287,]  0.0167484302  0.99795990 -4.154036e-04 -0.0577352391 -0.021494251
#> [288,]  0.0154519923  0.99874673  1.002603e-02 -0.0437559544 -0.015846293
#> [289,]  0.0233695755  0.99825105 -7.120332e-04 -0.0499267869 -0.021342605
#> [290,]  0.0038595502  0.99859970 -5.842400e-04 -0.0467801355 -0.024393225
#> [291,]  0.0206036799  0.99845610  5.832616e-03 -0.0471504370 -0.020092930
#> [292,]  0.0066254457  0.99839465 -7.128889e-03 -0.0495564855 -0.025642900
#> [293,]  0.0092068822  0.99836894 -3.691118e-03 -0.0537791919 -0.016397401
#> [294,]  0.0180222435  0.99848181  2.394844e-03 -0.0429277306 -0.029338429
#> [295,]  0.0115400962  0.99813572 -7.189130e-03 -0.0518059887 -0.029263602
#> [296,]  0.0156890295  0.99871502  5.892857e-03 -0.0449009338 -0.016472228
#> [297,]  0.0203778773  0.99865651  1.424225e-03 -0.0413328290 -0.023653319
#> [298,]  0.0068512483  0.99819424 -2.720498e-03 -0.0553740934 -0.022082511
#> [299,]  0.0159538098  0.99873845  6.728327e-03 -0.0444592958 -0.015655015
#> [300,]  0.0210872252  0.99808933 -4.808050e-03 -0.0534737643 -0.022146337
#> [301,]  0.0061419004  0.99876142  3.511777e-03 -0.0432331582 -0.023589493
#> [302,]  0.0114235336  0.99824171 -9.487334e-05 -0.0557607449 -0.016544279
#> [303,]  0.0158055920  0.99860903 -1.201400e-03 -0.0409461775 -0.029191551
#> [304,]  0.0054041076  0.99825971 -2.429454e-03 -0.0537280112 -0.023574151
#> [305,]  0.0218250180  0.99859104  1.133181e-03 -0.0429789113 -0.022161678
#> [306,]  0.0177024785  0.99803950 -3.287188e-03 -0.0565512174 -0.019873101
#> [307,]  0.0095266472  0.99881125  1.990915e-03 -0.0401557051 -0.025862729
#> [308,]  0.0114285388  0.99808619 -4.460702e-03 -0.0520541150 -0.031045471
#> [309,]  0.0158005869  0.99876456  3.164429e-03 -0.0446528074 -0.014690359
#> [310,]  0.0009472606  0.99890145  6.402902e-03 -0.0396701118 -0.024088834
#> [311,]  0.0124617181  0.99794789 -1.866760e-03 -0.0581723844 -0.023605026
#> [312,]  0.0147674076  0.99890286  5.704866e-04 -0.0385345380 -0.022130804
#> [313,]  0.0094168405  0.99842454  3.134993e-03 -0.0458810308 -0.030738798
#> [314,]  0.0178122852  0.99842621 -4.431267e-03 -0.0508258917 -0.014997031
#> [315,]  0.0161863487  0.99883782 -5.056498e-03 -0.0415067234 -0.017681297
#> [316,]  0.0110427770  0.99801293  3.760225e-03 -0.0552001991 -0.028054532
#> [317,]  0.0103329676  0.99821981 -3.899489e-03 -0.0493097515 -0.031682661
#> [318,]  0.0168961580  0.99863094  2.603216e-03 -0.0473971709 -0.014053168
#> [319,]  0.0073888524  0.99859194 -5.171132e-03 -0.0495701462 -0.016601206
#> [320,]  0.0198402732  0.99825881  3.874859e-03 -0.0471367762 -0.029134624
#> [321,]  0.0108973959  0.99798382  4.009294e-03 -0.0555867486 -0.028347436
#> [322,]  0.0224441698  0.99826300  1.267942e-03 -0.0508806311 -0.019410980
#> [323,]  0.0047849558  0.99858774 -2.564216e-03 -0.0458262913 -0.026324849
#> [324,]  0.0142343700  0.99894427  9.166290e-04 -0.0392112868 -0.019218886
#> [325,]  0.0129947557  0.99790648 -2.212902e-03 -0.0574956356 -0.026516944
#> [326,]  0.0122443039  0.99892321  3.635288e-03 -0.0406033340 -0.018456678
#> [327,]  0.0149848217  0.99792754 -4.931561e-03 -0.0561035884 -0.027279152
#> [328,]  0.0069175296  0.99886971 -3.879307e-03 -0.0420384704 -0.020716309
#> [329,]  0.0203115961  0.99798104  2.583034e-03 -0.0546684520 -0.025019520
#> [330,]  0.0141871222  0.99841413  2.889704e-03 -0.0524412092 -0.014474899
#> [331,]  0.0130420035  0.99843662 -4.185977e-03 -0.0442657133 -0.031260931
#> [332,]  0.0142885664  0.99839282  3.518167e-03 -0.0531664086 -0.012983548
#> [333,]  0.0043630175  0.99855197 -1.342048e-03 -0.0499327712 -0.019489481
#> [334,]  0.0228661082  0.99829878  4.577518e-05 -0.0467741513 -0.026246349
#> [335,]  0.0159728085  0.99845701 -1.490777e-03 -0.0431607365 -0.031038589
#> [336,]  0.0112563171  0.99839374  1.945038e-04 -0.0535461860 -0.014697241
#> [337,]  0.0150187927  0.99833476 -9.195688e-03 -0.0515357951 -0.019017109
#> [338,]  0.0122103330  0.99851599  7.899415e-03 -0.0451711273 -0.026718721
#> [339,]  0.0112945181  0.99847642  8.310014e-03 -0.0497407198 -0.019341088
#> [340,]  0.0159346076  0.99837433 -9.606287e-03 -0.0469662026 -0.026394742
#> [341,]  0.0157376221  0.99801450 -2.086534e-03 -0.0575753556 -0.020002929
#> [342,]  0.0114915036  0.99883625  7.902611e-04 -0.0391315669 -0.025732900
#> [343,]  0.0087062617  0.99835999  1.829477e-02 -0.0512785468 -0.015406160
#> [344,]  0.0155477353  0.99863616 -6.082129e-03 -0.0413767893 -0.027112070
#> [345,]  0.0116813904  0.99821459  4.785856e-03 -0.0553301331 -0.018623760
#> [346,]  0.0099783917  0.99864755 -7.364714e-03 -0.0429985642 -0.026464890
#> [347,]  0.0172507340  0.99820320  6.068441e-03 -0.0537083582 -0.019270940
#> [348,]  0.0072281142  0.99816124  8.311414e-04 -0.0558807616 -0.022327858
#> [349,]  0.0200010115  0.99868950 -2.127415e-03 -0.0408261609 -0.023407972
#> [350,]  0.0128734517  0.99885996  4.572956e-03 -0.0398813411 -0.022396828
#> [351,]  0.0143556740  0.99799079 -5.869230e-03 -0.0568255814 -0.023339002
#> [352,]  0.0136033520  0.99857373 -6.671215e-03 -0.0425977863 -0.028396863
#> [353,]  0.0136257736  0.99827702  5.374942e-03 -0.0541091362 -0.017338967
#> [354,]  0.0101134213  0.99793011  9.189008e-03 -0.0609777130 -0.015182184
#> [355,]  0.0106281339  0.99847112 -9.038769e-03 -0.0501188079 -0.018677849
#> [356,]  0.0166009917  0.99837962  7.742496e-03 -0.0465881146 -0.027057981
#> [357,]  0.0201683073  0.99851103  5.168000e-03 -0.0439227605 -0.024759795
#> [358,]  0.0070608184  0.99833972 -6.464273e-03 -0.0527841620 -0.020976034
#> [359,]  0.0196152276  0.99848575  3.786614e-03 -0.0433825537 -0.027296020
#> [360,]  0.0076138980  0.99836500 -5.082887e-03 -0.0533243687 -0.018439810
#> [361,]  0.0224498247  0.99821063  1.663873e-03 -0.0484074873 -0.026934903
#> [362,]  0.0047793010  0.99864012 -2.960146e-03 -0.0482994352 -0.018800927
#> [363,]  0.0192863019  0.99876646  2.022477e-03 -0.0412001273 -0.019800450
#> [364,]  0.0079428238  0.99808429 -3.318750e-03 -0.0555067951 -0.025935380
#> [365,]  0.0056956953  0.99862767 -2.720386e-03 -0.0483055134 -0.019222918
#> [366,]  0.0136345124  0.99796688  3.996491e-03 -0.0570009576 -0.024721099
#> [367,]  0.0135946132  0.99888387 -5.292764e-03 -0.0397059648 -0.021014730
#> [368,]  0.0181155726  0.99868066  5.915805e-03 -0.0423949161 -0.021827333
#> [369,]  0.0091135530  0.99817009 -7.212079e-03 -0.0543120063 -0.023908497
#> [370,]  0.0164705130  0.99853238  8.567217e-03 -0.0461128869 -0.021494653
#> [371,]  0.0107586127  0.99831837 -9.863490e-03 -0.0505940356 -0.024241177
#> [372,]  0.0151297195  0.99837807  5.445786e-03 -0.0455516523 -0.030128266
#> [373,]  0.0120994061  0.99847268 -6.742059e-03 -0.0511552701 -0.015607563
#> [374,]  0.0138077261  0.99798797 -6.540421e-03 -0.0558751926 -0.025778684
#> [375,]  0.0134213995  0.99886278  5.244147e-03 -0.0408317298 -0.019957145
#> [376,]  0.0173769543  0.99852961  1.149369e-02 -0.0453995852 -0.021057715
#> [377,]  0.0178253453  0.99864291 -8.040490e-03 -0.0433973133 -0.021133452
#> [378,]  0.0094037804  0.99820784  6.744217e-03 -0.0533096092 -0.024602378
#> [379,]  0.0158035521  0.99864484 -2.831814e-03 -0.0476605551 -0.013385918
#> [380,]  0.0114255736  0.99820591  1.535541e-03 -0.0490463674 -0.032349912
#> [381,]  0.0182883301  0.99877235 -6.425979e-04 -0.0432992876 -0.015623091
#> [382,]  0.0089407956  0.99807839 -6.536753e-04 -0.0534076348 -0.030112739
#> [383,]  0.0124454889  0.99895810 -2.391306e-03 -0.0390331942 -0.019962727
#> [384,]  0.0147836368  0.99789265  1.095033e-03 -0.0576737283 -0.025773103
#> [385,]  0.0155349833  0.99794573 -3.616815e-03 -0.0542502805 -0.030113125
#> [386,]  0.0116941424  0.99890501  2.320542e-03 -0.0424566419 -0.015622705
#> [387,]  0.0053677971  0.99785795  1.382640e-02 -0.0580508928 -0.026260804
#> [388,]  0.0219605028  0.99807069  1.827065e-03 -0.0532566591 -0.023087410
#> [389,]  0.0052686228  0.99878006 -3.123338e-03 -0.0434502633 -0.022648419
#> [390,]  0.0164429282  0.99850479  8.694843e-03 -0.0462174339 -0.022498006
#> [391,]  0.0107861975  0.99834596 -9.991116e-03 -0.0504894885 -0.023237824
#> [392,]  0.0153422435  0.99858845  2.448159e-03 -0.0415226599 -0.029250571
#> [393,]  0.0118868822  0.99826230 -3.744433e-03 -0.0551842626 -0.016485259
#> [394,]  0.0084855491  0.99822283 -7.364578e-03 -0.0520793006 -0.026696282
#> [395,]  0.0187435765  0.99862792  6.068305e-03 -0.0446276219 -0.019039548
#> [396,]  0.0218584803  0.99842445  1.601524e-03 -0.0485361268 -0.017677525
#> [397,]  0.0053706454  0.99842629 -2.897798e-03 -0.0481707956 -0.028058305
#> [398,]  0.0173366321  0.99849334  1.164870e-02 -0.0455403472 -0.022380219
#> [399,]  0.0163748150  0.99887241  2.563278e-03 -0.0398477499 -0.019782866
#> [400,]  0.0108543106  0.99797834 -3.859551e-03 -0.0568591725 -0.025952964
#> [401,]  0.0119373772  0.99884332  3.852921e-03 -0.0396783431 -0.024089324
#> [402,]  0.0152917485  0.99800742 -5.149194e-03 -0.0570285793 -0.021646506
#> [403,]  0.0062536901  0.99868158 -2.943619e-03 -0.0432856858 -0.026714601
#> [404,]  0.0209754355  0.99816916  1.647346e-03 -0.0534212367 -0.019021229
#> [405,]  0.0045019069  0.99860539  1.732517e-03 -0.0488063162 -0.019543584
#> [406,]  0.0227272188  0.99824536 -3.028790e-03 -0.0479006062 -0.026192246
#> [407,]  0.0141234426  0.99848076 -4.575939e-03 -0.0433675061 -0.030578120
#> [408,]  0.0131056831  0.99836999  3.279666e-03 -0.0533394164 -0.015157710
#> [409,] -0.0019864935  0.99863715  3.427311e-03 -0.0491240252 -0.017174864
#> [410,]  0.0202936597  0.99808068 -1.507574e-03 -0.0501971394 -0.030018421
#> [411,]  0.0069354660  0.99877007  2.113003e-04 -0.0465097831 -0.015717409
#> [412,]  0.0160790200  0.99787394 -4.929781e-04 -0.0572456741 -0.026678887
#> [413,]  0.0111501057  0.99897681 -8.032952e-04 -0.0394612484 -0.019056943
#> [414,]  0.0135182471  0.99860244 -9.357667e-03 -0.0468121064 -0.018206977
#> [415,]  0.0137108786  0.99824831  8.061394e-03 -0.0498948161 -0.027528852
#> [416,]  0.0204933789  0.99820259 -6.924218e-03 -0.0492668861 -0.026390042
#> [417,]  0.0067357468  0.99864816  5.627945e-03 -0.0474400364 -0.019345788
#> [418,]  0.0213787568  0.99832594 -6.806306e-03 -0.0476129455 -0.023978166
#> [419,]  0.0058503689  0.99852481  5.510033e-03 -0.0490939769 -0.021757664
#> [420,] -0.0025882454  0.99846541  1.220142e-02 -0.0498904886 -0.020547465
#> [421,]  0.0192158510  0.99833977  4.132426e-03 -0.0514519832 -0.016854572
#> [422,]  0.0080132747  0.99851098 -5.428699e-03 -0.0452549393 -0.028881258
#> [423,]  0.0191364222  0.99815362 -6.486972e-03 -0.0495808033 -0.028684574
#> [424,]  0.0080927035  0.99869713  5.190699e-03 -0.0471261191 -0.017051256
#> [425,]  0.0203593147  0.99834726 -7.521818e-03 -0.0491862055 -0.020307069
#> [426,]  0.0068698110  0.99850348  6.225544e-03 -0.0475207170 -0.025428761
#> [427,]  0.0085045196  0.99836396 -8.906521e-03 -0.0506519675 -0.023497450
#> [428,]  0.0187246061  0.99848679  7.610247e-03 -0.0460549549 -0.022238380
#> [429,]  0.0162047770  0.99841337  8.954372e-03 -0.0484878260 -0.021837828
#> [430,]  0.0110243486  0.99843738 -1.025065e-02 -0.0482190965 -0.023898002
#> [431,]  0.0031299733  0.99881126  1.043652e-02 -0.0460171521 -0.011822466
#> [432,]  0.3410058690 -0.11290652 -1.451996e-01  0.0357833010  0.921196912
#> [433,]  0.4702122875  0.23236804  4.978651e-01  0.3133039783  0.615529408
#> [434,] -0.0708170825 -0.37766936 -1.031701e-01  0.3478981202  0.848924985
#> [435,]  0.2323043784 -0.28975604  1.850005e-01  0.5184984385  0.747669901
#> [436,] -0.1021679275  0.48938265 -8.432954e-02  0.1108270564  0.854793676
#> [437,]  0.0418853529 -0.10728341  3.732179e-01  0.7647209024  0.512490234
#> [438,]  0.0716985858 -0.08321660 -4.229198e-01 -0.7321413489  0.522534417
#> [439,]  0.3410058690 -0.11290652 -1.451996e-01  0.0357833010  0.921196912
#> [440,]  0.3327294919 -0.11413803 -1.678788e-01 -0.0144788923  0.920798932
#> [441,]  0.3234180576 -0.11501448 -1.900357e-01 -0.0646960444  0.917536523
#> [442,]  0.3131005319 -0.11553314 -2.116015e-01 -0.1147119397  0.911419836
#> [443,]  0.3018090109 -0.11569240 -2.325091e-01 -0.1643709882  0.902467896
#> [444,]  0.2895786201 -0.11549176 -2.526933e-01 -0.2135187102  0.890708553
#> [445,]  0.2764474061 -0.11493185 -2.720915e-01 -0.2620022167  0.876178387
#> [446,]  0.2624562175 -0.11401441 -2.906432e-01 -0.3096706848  0.858922599
#> [447,]  0.2476485781 -0.11274229 -3.082908e-01 -0.3563758270  0.838994869
#> [448,]  0.2320705517 -0.11111945 -3.249794e-01 -0.4019723527  0.816457187
#> [449,]  0.2157705985 -0.10915094 -3.406570e-01 -0.4463184198  0.791379664
#> [450,]  0.1987994246 -0.10684288 -3.552749e-01 -0.4892760763  0.763840311
#> [451,]  0.1812098240 -0.10420246 -3.687877e-01 -0.5307116893  0.733924799
#> [452,] -0.2511047687  0.81571089  4.529878e-02 -0.1976382834  0.480051323
#> [453,]  0.1812098240 -0.10420246 -3.687877e-01 -0.5307116893  0.733924799
#> [454,]  0.1607502552 -0.04850633 -3.576274e-01 -0.5318506519  0.749028735
#> [455,]  0.1396678384  0.00737774 -3.450814e-01 -0.5309288895  0.761230462
#> [456,]  0.1180442603  0.06323323 -3.311983e-01 -0.5279499735  0.770482701
#> [457,]  0.0959633043  0.11884371 -3.160320e-01 -0.5229254462  0.776749604
#> [458,]  0.0735105258  0.17399371 -2.996412e-01 -0.5158747756  0.780006889
#> [459,]  0.0507729212  0.22846956 -2.820894e-01 -0.5068252806  0.780241935
#> [460,]  0.0278385902  0.28206016 -2.634446e-01 -0.4958120246  0.777453832
#> [461,]  0.0047963949  0.33455789 -2.437790e-01 -0.4828776799  0.771653381
#> [462,] -0.0182643846  0.38575933 -2.231689e-01 -0.4680723623  0.762863059
#> [463,] -0.0412543963  0.43546610 -2.016941e-01 -0.4514534370  0.751116923
#> [464,] -0.0640845625  0.48348560 -1.794378e-01 -0.4330852962  0.736460487
#> [465,] -0.0866664246  0.52963176 -1.564862e-01 -0.4130391096  0.718950538
#> [466,] -0.1089124864  0.57372580 -1.329284e-01 -0.3913925488  0.698654920
#> [467,] -0.1307365524  0.61559686 -1.088554e-01 -0.3682294863  0.675652273
#> [468,] -0.1520540625  0.65508272 -8.436073e-02 -0.3436396704  0.650031722
#> [469,] -0.1727824192  0.69203037 -5.953915e-02 -0.3177183775  0.621892538
#> [470,] -0.1928413078  0.72629665 -3.448689e-02 -0.2905660431  0.591343750
#> [471,] -0.2121530074  0.75774881 -9.301000e-03 -0.2622878725  0.558503723
#> [472,] -0.2306426924  0.78626497  1.592093e-02 -0.2329934332  0.523499701
#> [473,] -0.2482387221  0.81173464  4.108117e-02 -0.2027962302  0.486467310
#> [474,] -0.2511047687  0.81571089  4.529878e-02 -0.1976382834  0.480051323
#> [475,] -0.4655322819  0.73361103  7.709151e-02 -0.3828522256  0.304262418
#> [476,] -0.7117655922  0.45356251 -1.782805e-02  0.4999175875 -0.193482182
#> [477,] -0.5712587214  0.41767430  1.036473e-01  0.2082586362  0.667156078
#> [478,] -0.4143199215  0.23672482 -7.147369e-01 -0.3187152526  0.399840191
#> [479,] -0.4154921951  0.57769950  7.742047e-02 -0.4672418849  0.518961097
#> [480,] -0.4580949190  0.55089868  6.242979e-02 -0.6379167961 -0.275362253
#> [481,]  0.1840226289  0.84211555 -5.046149e-01 -0.0313577375 -0.036845854
#> [482,] -0.2511047687  0.81571089  4.529878e-02 -0.1976382834  0.480051323
#> [483,] -0.2276957401  0.83830414  8.782653e-03 -0.1912835457  0.456874479
#> [484,] -0.2036105885  0.85840812 -2.775955e-02 -0.1843608080  0.432340986
#> [485,] -0.1789208329  0.87596313 -6.421933e-02 -0.1768906267  0.406523694
#> [486,] -0.1536997873  0.89091704 -1.004884e-01 -0.1688951839  0.379499264
#> [487,] -0.1280223433  0.90322546 -1.364591e-01 -0.1603982214  0.351347944
#> [488,] -0.1019647481  0.91285182 -1.720246e-01 -0.1514249702  0.322153327
#> [489,] -0.0756043773  0.91976755 -2.070793e-01 -0.1420020756  0.292002103
#> [490,] -0.0490195058  0.92395210 -2.415190e-01 -0.1321575180  0.260983803
#> [491,] -0.0222890751  0.92539306 -2.752416e-01 -0.1219205301  0.229190535
#> [492,]  0.2550251841  0.59934672 -2.261639e-01 -0.5568030319 -0.463212617
#> [493,]  0.3725019069  0.50848894 -7.526268e-01  0.0764889974 -0.174308996
#> [494,]  0.0115187504  0.97737315  6.514837e-02  0.2009071277  0.001032177
#> [495,] -0.0222890751  0.92539306 -2.752416e-01 -0.1219205301  0.229190535
#> [496,] -0.0192658495  0.94177875 -2.454243e-01 -0.0919400708  0.209751140
#> [497,] -0.0161955493  0.95586328 -2.150073e-01 -0.0617349631  0.189799235
#> [498,] -0.0130856766  0.96761223 -1.840649e-01 -0.0313790109  0.169383570
#> [499,] -0.0099438300  0.97699689 -1.526728e-01 -0.0009463865  0.148554029
#> [500,] -0.0067776864  0.98399434 -1.209077e-01  0.0294885503  0.127361509
#> [501,] -0.0035949821  0.98858748 -8.884710e-02  0.0598514342  0.105857790
#> [502,] -0.0004034938  0.99076508 -5.656943e-02  0.0900680757  0.084095416
#> [503,] -0.1592005971  0.93351757  1.505059e-01 -0.0620697275  0.276939423
#> [504,]  0.3071423425  0.78233842  5.262080e-01  0.0806559140  0.101044282
#> [505,] -0.2247220473  0.84263266 -3.498804e-01  0.2037854278  0.274818792
#> [506,]  0.4586909765  0.45037070 -6.011041e-02 -0.3538596052  0.676711857
#> [507,] -0.5310439602  0.79923830 -8.970107e-02 -0.2430922323 -0.109865118
#> [508,] -0.3377467332  0.46533340  7.837140e-01 -0.2304129828 -0.045761587
#> [509,]  0.3068675670  0.53561719 -6.442545e-01 -0.4430146135 -0.087296594
#> [510,]  0.4183830894  0.69438038  1.476780e-01 -0.5016832308  0.263242547
#> [511,]  0.0216983458  0.52662868  3.410311e-01  0.3479926704 -0.696268861
#> [512,]  0.3096808220  0.69334367  5.988242e-01  0.1318540791 -0.217707141
#> [513,]  0.5117720161  0.69053365 -1.309114e-01  0.3366386611  0.361648043
#> [514,] -0.2057703256  0.79664412 -5.473783e-01 -0.0359103426  0.148674616
#> [515,]  0.1183250376  0.95554145 -2.548009e-01  0.0840921243 -0.030736990
#> [516,]  0.4340757088  0.57457155  4.791157e-01  0.3659333222  0.343492058
#> [517,] -0.1000386856  0.75310158  2.915911e-03 -0.6491285906 -0.038129374
#> [518,] -0.5837540442  0.45742425  1.825436e-01 -0.6421227194 -0.065958545
#> [519,] -0.3016186270  0.89976437  1.550154e-01 -0.0132590457  0.274307659
#> [520,] -0.1019447978  0.69163684 -1.677674e-01  0.6576482255  0.224941441
#> [521,] -0.3265397072  0.68824476  5.361298e-01  0.2118324954  0.295605786
#> [522,]  0.6308744721  0.39504801 -3.986774e-01 -0.2466388006  0.475562923
#> [523,] -0.6444495734  0.63264163  3.965137e-01  0.1472234535 -0.074507969
#> [524,]  0.2276934292  0.90755709 -1.730098e-02 -0.1646844562 -0.311569483
#> [525,] -0.3077582205  0.66539709  1.569924e-01  0.2741682067  0.602259719
#> [526,]  0.4420953896  0.42774572  2.213452e-01  0.2262441477  0.722083893
#> 
#> $flip
#> [1] TRUE
#> 
#> $dt
#> # A tibble: 526 x 8
#>    basis          index_val tries info             loop method       alpha    id
#>    <list>             <dbl> <dbl> <chr>           <dbl> <chr>        <dbl> <int>
#>  1 <dbl[,1] [5 ×…     0.749     1 new_basis           1 search_geod…   0.5     1
#>  2 <dbl[,1] [5 ×…     0.749     2 direction_sear…     1 search_geod…  NA       2
#>  3 <dbl[,1] [5 ×…     0.749     2 direction_sear…     1 search_geod…  NA       3
#>  4 <dbl[,1] [5 ×…     0.749     2 direction_sear…     1 search_geod…  NA       4
#>  5 <dbl[,1] [5 ×…     0.749     2 direction_sear…     1 search_geod…  NA       5
#>  6 <dbl[,1] [5 ×…     0.749     2 direction_sear…     1 search_geod…  NA       6
#>  7 <dbl[,1] [5 ×…     0.749     2 direction_sear…     1 search_geod…  NA       7
#>  8 <dbl[,1] [5 ×…     0.749     2 direction_sear…     1 search_geod…  NA       8
#>  9 <dbl[,1] [5 ×…     0.749     2 direction_sear…     1 search_geod…  NA       9
#> 10 <dbl[,1] [5 ×…     0.749     2 best_direction…     1 search_geod…  NA      10
#> # … with 516 more rows
#>