Extract directional search points during the optimisation

Extract directional search points during the optimisation

get_dir_search(dt, ratio = 5, ...)

Arguments

dt
a data object collected by the projection pursuit guided tour optimisation in the tourr package

ratio
numeric; a buffer value to deviate directional search points from the anchor points

...
arguments passed to compute_pca()

Value

a tibble object containing the directional search bases in pseudo derivative search

See also

Examples

holes_1d_geo %>%
  compute_pca() %>%
  purrr::pluck("aug") %>%
  get_dir_search()

#> # A tibble: 330 x 16
#>    basis index_val info  method alpha tries  loop    id row_num   PC1    PC2
#>    <lis>     <dbl> <chr> <chr>  <dbl> <dbl> <dbl> <dbl>   <int> <dbl>  <dbl>
#>  1 <dbl…     0.749 dire… PD        NA     2     1     2       2 -1.80 -0.776
#>  2 <dbl…     0.749 dire… PD        NA     2     1     3       3 -1.77 -0.601
#>  3 <dbl…     0.749 dire… PD        NA     2     1     4       4 -1.79 -0.632
#>  4 <dbl…     0.749 dire… PD        NA     2     1     5       5 -1.78 -0.745
#>  5 <dbl…     0.749 dire… PD        NA     2     1     6       6 -1.80 -0.562
#>  6 <dbl…     0.749 dire… PD        NA     2     1     7       7 -1.77 -0.815
#>  7 <dbl…     0.749 dire… PD        NA     2     1     8       8 -1.73 -0.732
#>  8 <dbl…     0.749 dire… PD        NA     2     1     9       9 -1.84 -0.644
#>  9 <dbl…     0.749 best… PD        NA     2     1    10      10 -1.78 -0.804
#> 10 <dbl…     0.749 dire… PD        NA     2     1    11      11 -1.79 -0.573
#> # … with 320 more rows, and 5 more variables: PC3 <dbl>, PC4 <dbl>, PC5 <dbl>,
#> #   anchor_x <dbl>, anchor_y <dbl>