Function to calculate smoothness and squintability

Function to calculate smoothness and squintability

calc_smoothness(
  idx,
  data = sine1000,
  n_basis = 300,
  n = 6,
  d = 2,
  best = matrix(c(0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1), nrow = 6),
  start_parms = c(0.001, 0.5, 2, 2),
  other_gp_params = NULL
)

calc_squintability(
  idx,
  data = sine1000,
  method = c("ks", "nls"),
  n_basis = 100,
  n = 6,
  d = 2,
  proj_dist_threshold = 1.5,
  step = 0.02,
  best = matrix(c(0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1), nrow = 6),
  bin_nobs_threshold = 5000,
  bin_width = 0.02,
  sampling_seed = 123
)

Arguments

idx

character, the name of projection pursuit index function, e.g. "holes" (see the tourr package for index examples)

data

matrix, the data to be projected

n_basis

numeric, the number of random basis to generate for calculating smoothness

n

dimensionality of data

d

dimensionality of target projection

best

a matrix, the theoretical best projection matrix, used to calculate projection distance with the simulated random bases.

start_parms

Optionally specified starting values for parameters. If NULL, fit_model will select default starting values.

other_gp_params

list, other parameters to be passed to GpGp::fit_model

method

only for squintability, the method to calculate squintability, either through kernel smoothing ("ks") or non-linear least square ("nls")

proj_dist_threshold

only for squintability, the threshold for projection distance for the random basis to be considered in sampling

step

only for squintability, the step size for interpolation, recommend between 0.01 and 0.1

bin_nobs_threshold

numeric, only for squintability, the threshold number of observations for applying binning before fitting the kernel

bin_width

only for squintability, the bin size for binning the data before fitting the kernel

sampling_seed

the seed used for sampling the random basis