Bind random bases in the projection bases space

Given the orthonormality constraint, the projection bases live in a high dimensional hollow sphere. Generating random points on the sphere is useful to preceive the data object in the high dimensional space.

bind_random(dt, n = 500)

Arguments

dt
A data object from the running the optimisation algorithm in guided tour

n
Number of random points to generate in each dimension

See also

Examples

bind_random(holes_1d_better) %>% tail(5)

#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
#> Using compatibility `.name_repair`.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> # A tibble: 5 x 8
#>   basis          index_val tries info            loop method         alpha    id
#>   <list>             <dbl> <dbl> <chr>          <dbl> <chr>          <dbl> <dbl>
#> 1 <dbl[,1] [5 ×…        NA    NA randomly_gene…    NA randomly_gene…    NA     0
#> 2 <dbl[,1] [5 ×…        NA    NA randomly_gene…    NA randomly_gene…    NA     0
#> 3 <dbl[,1] [5 ×…        NA    NA randomly_gene…    NA randomly_gene…    NA     0
#> 4 <dbl[,1] [5 ×…        NA    NA randomly_gene…    NA randomly_gene…    NA     0
#> 5 <dbl[,1] [5 ×…        NA    NA randomly_gene…    NA randomly_gene…    NA     0