Bind random bases in the projection bases space as a matrix

Bind random bases in the projection bases space as a matrix

bind_random_matrix(basis, n = 500, front = FALSE)

Arguments

basis
A basis matrix returned by get_basis_matrix()

n
Number of random points to generate in each dimension

front
if the random matrix should be bound before or after the data basis

Value

matrix

See also

Examples

data <- get_basis_matrix(holes_1d_geo)
bind_random_matrix(data) %>% tail(5)

#>                  V1           V2         V3        V4         V5
#> [2912,] -0.11500792  0.108791393 -0.2119039 0.6421904  0.7194622
#> [2913,] -0.64030197  0.470754738  0.3598407 0.2095701  0.4415862
#> [2914,]  0.08192813 -0.400575298  0.2188422 0.2890079 -0.8375021
#> [2915,] -0.77069449 -0.002085302  0.2361071 0.5335136  0.2562077
#> [2916,] -0.53992262  0.069524169  0.4387449 0.1042930 -0.7073018