Compute PCA for the projection bases

Compute PCA for the projection bases

compute_pca(dt, group = NULL, random = TRUE)

Arguments

dt
A data object from the running the optimisation algorithm in guided tour

group
The grouping variable, useful when there are multiple algorithms in the data object

random
Boolean, if the random data from the high dimensional sphere should be bounded

Examples

dplyr::bind_rows(holes_1d_geo, holes_1d_better) %>% compute_pca(group = method)

#> signs in all the bases will be fliped in group search_geodesic

#> $pca_summary
#> Standard deviations (1, .., p=5):
#> [1] 1.0464239 1.0064726 0.9903015 0.9888237 0.9661989
#> 
#> Rotation (n x k) = (5 x 5):
#>            PC1        PC2        PC3         PC4         PC5
#> V1  0.50410202 -0.3520146  0.1909697 -0.55986696 -0.52158069
#> V2 -0.40303443  0.1969047  0.8832958 -0.09016991 -0.10222380
#> V3 -0.46188590 -0.4268000 -0.1256141 -0.61624531  0.45712799
#> V4  0.04286299 -0.7945427  0.2627765  0.53693787  0.09752381
#> V5  0.60685136  0.1544600  0.3138292 -0.10177334  0.70641842
#> 
#> $aug
#> # A tibble: 3,012 x 14
#>    basis index_val tries info   loop method alpha    id row_num   PC1    PC2
#>    <lis>     <dbl> <dbl> <chr> <dbl> <chr>  <dbl> <dbl>   <int> <dbl>  <dbl>
#>  1 <dbl…     0.749     1 new_…     1 searc…   0.5     1       1 -1.82 -0.146
#>  2 <dbl…     0.749     2 dire…     1 searc…  NA       2       2 -1.83 -0.162
#>  3 <dbl…     0.749     2 dire…     1 searc…  NA       3       3 -1.81 -0.130
#>  4 <dbl…     0.749     2 dire…     1 searc…  NA       4       4 -1.82 -0.138
#>  5 <dbl…     0.749     2 dire…     1 searc…  NA       5       5 -1.82 -0.153
#>  6 <dbl…     0.749     2 dire…     1 searc…  NA       6       6 -1.82 -0.125
#>  7 <dbl…     0.749     2 dire…     1 searc…  NA       7       7 -1.82 -0.167
#>  8 <dbl…     0.749     2 dire…     1 searc…  NA       8       8 -1.81 -0.151
#>  9 <dbl…     0.749     2 dire…     1 searc…  NA       9       9 -1.83 -0.140
#> 10 <dbl…     0.749     2 best…     1 searc…  NA      10      10 -1.82 -0.166
#> # … with 3,002 more rows, and 3 more variables: PC3 <dbl>, PC4 <dbl>, PC5 <dbl>
#>