Compute PCA for the projection bases

Compute PCA for the projection bases

compute_pca(dt, group = NULL, random = TRUE)

Arguments

dt
A data object from the running the optimisation algorithm in guided tour

group
The grouping variable, useful when there are multiple algorithms in the data object

random
Boolean, if the random data from the high dimensional sphere should be bounded

Examples

dplyr::bind_rows(holes_1d_geo, holes_1d_better) %>% compute_pca(group = method)

#> $pca_summary
#> Standard deviations (1, .., p=5):
#> [1] 1.0227646 1.0093060 0.9979284 0.9926170 0.9767827
#> 
#> Rotation (n x k) = (5 x 5):
#>            PC1         PC2        PC3       PC4        PC5
#> V1  0.42462146 -0.46965525 -0.5808180 0.1775849  0.4798277
#> V2 -0.56927887 -0.35056714  0.4119725 0.1855313  0.5906616
#> V3 -0.53190340 -0.09410759 -0.4800363 0.5573989 -0.4087715
#> V4 -0.01982495  0.80132647 -0.1716976 0.3178762  0.4764006
#> V5  0.46076451 -0.07447332  0.4827155 0.7227053 -0.1638067
#> 
#> $aug
#> # A tibble: 5,526 x 15
#>    basis index_val tries info   loop method alpha    id row_num   PC1   PC2
#>    <lis>     <dbl> <dbl> <chr> <dbl> <chr>  <dbl> <dbl>   <int> <dbl> <dbl>
#>  1 <dbl…     0.749     1 new_…     1 searc…   0.5     1       1 -1.53 0.392
#>  2 <dbl…     0.749     2 dire…     1 searc…  NA       2       2 -1.54 0.391
#>  3 <dbl…     0.749     2 dire…     1 searc…  NA       3       3 -1.52 0.393
#>  4 <dbl…     0.749     2 dire…     1 searc…  NA       4       4 -1.53 0.371
#>  5 <dbl…     0.749     2 dire…     1 searc…  NA       5       5 -1.53 0.413
#>  6 <dbl…     0.749     2 dire…     1 searc…  NA       6       6 -1.53 0.375
#>  7 <dbl…     0.749     2 dire…     1 searc…  NA       7       7 -1.54 0.409
#>  8 <dbl…     0.749     2 dire…     1 searc…  NA       8       8 -1.52 0.391
#>  9 <dbl…     0.749     2 dire…     1 searc…  NA       9       9 -1.54 0.393
#> 10 <dbl…     0.749     2 best…     1 searc…  NA      10      10 -1.54 0.409
#> # … with 5,516 more rows, and 4 more variables: PC3 <dbl>, PC4 <dbl>,
#> #   PC5 <dbl>, animate_id <int>
#>